refactor: Remove consensus peer endpoints (#17)
