+++ api/rest/route/healthcheck.go
@@ -1,7 +1,9 @@
 package route
 
 import (
+	"fmt"
 	"github.com/gofiber/fiber/v2"
+	"nubedb/api/rest/jsonresponse"
 )
 
 func (a *ApiCtx) healthCheck(fiberCtx *fiber.Ctx) error {
@@ -9,4 +11,11 @@ func (a *ApiCtx) healthCheck(fiberCtx *fiber.Ctx) error {
 		return fiberCtx.Status(fiber.StatusInternalServerError).SendString("")
 	}
 	return fiberCtx.Status(fiber.StatusOK).SendString("")
+}
+
+func (a *ApiCtx) consensusState(fiberCtx *fiber.Ctx) error {
+	stats := a.Node.Consensus.Stats()
+	address, id := a.Node.Consensus.LeaderWithID()
+	stats["leader"] = fmt.Sprintf("Address: %s Leader ID: %s", address, id)
+	return jsonresponse.OK(fiberCtx, "consensus state retrieved successfully", stats)
 }
